@@ -37,9 +37,8 @@
 		"svelte": "^3.0.0"
 	},
 	"devDependencies": {
-		"esm": "^3.2.25",
-		"mocha": "^7.2.0",
-		"svelte": "^3.24.0"
+		"mocha": "^9.1.3",
+		"svelte": "^3.44.1"
 	},
 	"license": "MIT"
 }
